requirements.txt
blockchain-etl==0.0.5
Flask==1.0.2
Flask-Cors==3.0.6
gunicorn==19.9.0
requests==2.19.1
switcheo==0.2.2
